pyproject.toml
[tool.poetry]
name = "todoist-ics"
version = "0.1.0"
description = "Create a duration-aware ics from scheduled Todoist tasks."
authors = ["Michel Thomas Brodatzki <michel@brodatzki.com>"]

[tool.poetry.dependencies]
python = "^3.9"
isodate = "^0.6.1"
ics = "^0.7.2"
python-dotenv = "^0.20.0"
bottle = "^0.12.23"
pytz = "^2022.2.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
todoist-ics = "todoist_ics:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
